pyproject.toml
[tool.poetry]
name = "rockband-app"
version = "0.1.0"
description = ""
authors = ["Julie Park <juliepark242@gmail.com>"]

[tool.poetry.dependencies]
python = "3.9.12"
Django = "^4.0.4"
psycopg2-binary = "^2.9.3"
djangorestframework = "^3.13.1"
django-cors-headers = "^3.11.0"
dj-rest-auth = "^2.2.4"
djangorestframework-simplejwt = "^5.1.0"
dj-database-url = "^0.5.0"
gunicorn = "^20.1.0"
whitenoise = "^6.0.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
mypy = "^0.942"
autopep8 = "^1.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
